// src/app/interfaces/interfaces.ts
// Generated by https://quicktype.io

export interface Campain {
    code:         number;
    title:        string;
    sender:       string;
    list_type:    string;
    initial_date: string;
    final_date:   string;
    total_msg:    number;
    msg_sent:     number;
    state:        string;
}